CLI scan-sources: no disabled by default
Correct the line removed in 8bb10248a17697700d730193003426b983485cd7

We do in fact want "load disabled" to default to False. But it should
also respect the --load-disabled option on the runtime.

Setting OPTS["disabled"] unconditionally to False as before would fix
the current problem (scans including disabled). But it wouldn't respect
the runtime options (never did). By removing it entirely from the OPTS,
the runtime governs again (and its default is what we want).